Add more tests showcasing the problem + move all GenericEventListenerTest tests in one test
