+++ modules/flowable-cmmn-engine/src/test/resources/org/flowable/cmmn/test/eventlistener/TimerEventListenerTest.testTimerExpressionDuration.cmmn
@@ -3,7 +3,7 @@
   <case id="testTimerExpression" name="testTimerExpression" flowable:initiatorVariableName="initiator">
     <casePlanModel id="casePlanModel">
       <planItem id="planItem1" definitionRef="timerListener"></planItem>
-      <planItem id="planItem2" name="A" definitionRef="sid-E85224AB-66DD-4C8C-8809-4F50DC117067">
+      <planItem id="planItem2" name="A" definitionRef="taskA">
         <entryCriterion id="sid-D752F3E8-D736-4A1C-9C65-FF012CBB47AA" sentryRef="sentry1"></entryCriterion>
       </planItem>
       <sentry id="sentry1">
@@ -14,7 +14,7 @@
       <timerEventListener id="timerListener" name="Timer listener">
         <timerExpression><![CDATA[PT1H]]></timerExpression>
       </timerEventListener>
-      <humanTask id="sid-E85224AB-66DD-4C8C-8809-4F50DC117067" name="A"></humanTask>
+      <humanTask id="taskA" name="A"></humanTask>
     </casePlanModel>
   </case>
   <cmmndi:CMMNDI>
